스크랩, 댓글 요청이 완료되지 않으면 또 다시 요청을 보내지 않도록 구현  (#268)
* feat: 댓글 수정 여부를 클라이언트에서 판단하는 기능 제거

* feat: 새 댓글이 리스트의 마지막에 등록되도록 수정

* feat: 댓글 요청이 완료되지 않으면 또 다시 요청 보내지 않도록 수정

* feat: Loading State 상태 하나로 통일

* fix: 댓글 삭제하고 재등록시 등록 안되는 버그 수정

+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/CommentViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.digginroom.digginroom.feature.room.customview.CommentState
 import com.digginroom.digginroom.livedata.NonNullMutableLiveData
 import com.digginroom.digginroom.model.CommentModel
 import com.digginroom.digginroom.model.mapper.CommentMapper.toModel
@@ -15,7 +14,9 @@ class CommentViewModel(
     private val commentRepository: CommentRepository
 ) : ViewModel() {
 
-    private val _comments: MutableLiveData<List<CommentModel>> = MutableLiveData()
+    private val _comments: NonNullMutableLiveData<List<CommentModel>> = NonNullMutableLiveData(
+        emptyList()
+    )
     val comments: LiveData<List<CommentModel>>
         get() = _comments
 
@@ -26,7 +27,7 @@ class CommentViewModel(
 
     fun findComments(roomId: Long) {
         writtenComment.value = ""
-        _commentState.value = CommentState.Create.Ready
+        _commentState.value = CommentState.Post.Ready
         viewModelScope.launch {
             commentRepository.findComments(roomId).onSuccess { comments ->
                 _comments.value = comments.map { it.toModel() }
@@ -35,41 +36,43 @@ class CommentViewModel(
     }
 
     fun postComment(roomId: Long, comment: String) {
+        if (_commentState.value == CommentState.Post.Loading) return
+        _commentState.value = CommentState.Post.Loading
+
         viewModelScope.launch {
             commentRepository.postComment(roomId, comment).onSuccess {
-                val oldComments: MutableList<CommentModel> =
-                    _comments.value?.toMutableList() ?: mutableListOf()
-                oldComments.add(0, it.toModel())
-                _comments.value = oldComments
-                _commentState.value = CommentState.Create.Succeed
+                _comments.value = _comments.value + it.toModel()
+                _commentState.value = CommentState.Post.Succeed
             }.onFailure {
-                _commentState.value = CommentState.Create.Failed
+                _commentState.value = CommentState.Post.Failed
             }
         }
     }
 
     fun updateWrittenComment(roomId: Long, commentId: Long, comment: String, updatePosition: Int) {
+        if (_commentState.value == CommentState.Update.Loading) return
+        _commentState.value = CommentState.Update.Loading
+
         viewModelScope.launch {
             commentRepository.updateComment(roomId, commentId, comment).onSuccess {
-                val oldComments: MutableList<CommentModel> =
-                    _comments.value?.toMutableList() ?: mutableListOf()
-                oldComments[updatePosition] =
-                    oldComments[updatePosition].copy(comment = comment, isUpdated = true)
-                _comments.value = oldComments
-                _commentState.value = CommentState.Edit.Succeed
+                _comments.value = _comments.value.toMutableList().apply {
+                    this[updatePosition] = this[updatePosition].copy(comment = comment)
+                }
+                _commentState.value = CommentState.Update.Succeed
             }.onFailure {
-                _commentState.value = CommentState.Edit.Failed
+                _commentState.value = CommentState.Update.Failed
             }
         }
     }
 
     fun deleteComment(roomId: Long, commentId: Long, deletePosition: Int) {
+        if (_commentState.value == CommentState.Delete.Loading) return
+        _commentState.value = CommentState.Delete.Loading
+
         viewModelScope.launch {
             commentRepository.deleteComment(roomId, commentId).onSuccess {
-                val oldComments: MutableList<CommentModel> =
-                    _comments.value?.toMutableList() ?: mutableListOf()
-                oldComments.removeAt(deletePosition)
-                _comments.value = oldComments
+                _comments.value =
+                    _comments.value.filterIndexed { index, _ -> index != deletePosition }
                 _commentState.value = CommentState.Delete.Succeed
             }.onFailure {
                 _commentState.value = CommentState.Delete.Failed
